+++ core/Material.cpp
@@ -20,20 +20,20 @@ Material* Material::createMaterial(Value& matSpec) {
     double specular = 0;
 
     if (matSpec.HasMember("ka") && matSpec["ka"].IsNumber()) {
-        ka = matSpec["ka"].GetFloat();
+        ka = matSpec["ka"].GetDouble();
     }
 
     if (matSpec.HasMember("ks") && matSpec["ks"].IsNumber()) {
-        ks = matSpec["ks"].GetFloat();
+        ks = matSpec["ks"].GetDouble();
     }
 
     if (matSpec.HasMember("kd") && matSpec["kd"].IsNumber()) {
-        kd = matSpec["kd"].GetFloat();
+        kd = matSpec["kd"].GetDouble();
     }
 
     if (matSpec.HasMember("specularexponent") &&
         matSpec["specularexponent"].IsNumber()) {
-        specular = matSpec["specularexponent"].GetFloat();
+        specular = matSpec["specularexponent"].GetDouble();
     }
 
     // Currently only BlinnPhong
@@ -42,20 +42,20 @@ Material* Material::createMaterial(Value& matSpec) {
     if (matSpec["diffusecolor"].IsArray() &&
         matSpec["diffusecolor"].Size() >= 3) {
         auto diffuseArray = matSpec["diffusecolor"].GetArray();
-        newMat->diffuse = Vec3d(diffuseArray[0].GetFloat(), diffuseArray[1].GetFloat(),
-                                 diffuseArray[2].GetFloat());
+        newMat->diffuse = Vec3d(diffuseArray[0].GetDouble(), diffuseArray[1].GetDouble(),
+                                 diffuseArray[2].GetDouble());
     }
 
     if (matSpec.HasMember("metallic") && matSpec["metallic"].IsNumber()) {
-        newMat->metallic = matSpec["metallic"].GetFloat();
+        newMat->metallic = matSpec["metallic"].GetDouble();
     }
 
     if (matSpec.HasMember("roughness") && matSpec["roughness"].IsNumber()) {
-        newMat->roughness = matSpec["roughness"].GetFloat();
+        newMat->roughness = matSpec["roughness"].GetDouble();
     }
 
     if (matSpec.HasMember("transmit") && matSpec["transmit"].IsNumber()) {
-        newMat->transmit = matSpec["transmit"].GetFloat();
+        newMat->transmit = matSpec["transmit"].GetDouble();
     }
 
     return newMat;
